bblayers.conf: Add template layer
The template layer is used as an example layer that provides a basic
layer with template structure, configuration and template component
recipes.

The layer is added to the PELUX image because the example library
and service will be later used to in the software factory to present
the recommend workflow on various platform and service development
topics.

Signed-off-by: Gordan Markuš <gordan.markus@pelagicore.com>